Moved types+interfaces to ../api and indexed-db to ../indexeddb-typings
