+++ api/dexie-internal.d.ts
