+++ api/promise-extended.d.ts
