+++ api/table.d.ts
