+++ api/then-shortcut.d.ts
