+++ api/transaction.d.ts
